test(onboard): move installer restore-intent caller tests to a new file
